s complete, percent complete

+++ utility/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: fholwerd <fholwerd@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/04/22 14:01:21 by fholwerd      #+#    #+#                 */
-/*   Updated: 2021/04/29 13:16:09 by fholwerd      ########   odam.nl         */
+/*   Updated: 2021/05/11 12:33:02 by fholwerd      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../header.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i] != '\0')
